Rename Request to CurlRequest
