+++ web/conw/lib/client/IpInfo.php
@@ -2,12 +2,12 @@
 
 namespace Client;
 
-require __DIR__ . "/../req/Request.php";
+require __DIR__ . "/../curl/CurlRequest.php";
 
 abstract class IpInfo {
   private static function fetch(string $ip): array {
     $url = "http://ip-api.com/json/$ip?fields=66846719";
-    $request = new \Req\Request($url);
+    $request = new \Curl\CurlRequest($url);
     return $request->json();
   }
